test_eval

+++ chromosome.h
@@ -8,7 +8,7 @@ class chromosome
 {
 public:
     std::vector<time_table> schedule;
-    double fitness;
+    int fitness;
 
 public:
     chromosome();
